refactor(BIP44Worker): split functions into individual files
